% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Generate denisty plot for some matrix statistic}
\usage{
densityPlot(mcmc_stats, variable, xlab)
}
\arguments{
\item{mcmc_stats}{list of outputs from MeanMatrixStatistics}

\item{variable}{variable to be ploted, any column in the mcmc_stats objects}

\item{xlab}{x axis label}
}
\value{
ggplot2 density plot object
}
\description{
Generate denisty plot for some matrix statistic
}
\examples{
data(P_stats)
r2_plot = densityPlot(P_stats, "MeanSquaredCorrelation",
                      "Mean squared correlation")
pc1.percent_plot = densityPlot(P_stats, "pc1.percent",
                               "Proportion of variation in E1") +
                     theme(legend.position = "none")
flexibility_plot = densityPlot(P_stats, "flexibility",
                               "Mean flexibility") +
                     theme(legend.position = "none")
evolvability_plot = densityPlot(P_stats, "evolvability",
                                "Mean evolvability") +
                      theme(legend.position = "none")
cond_evolvability_plot = densityPlot(P_stats, "conditional.evolvability",
                                     "Mean conditional evolvability") +
                           theme(legend.position = "none")
figure_3 <- ggdraw() +
    draw_plot(r2_plot, 0, 0.5, 0.5, 0.5) +
    draw_plot(pc1.percent_plot, 0.5, 0.5, 0.5, 0.5) +
    draw_plot(flexibility_plot, 0, 0, 0.5, 0.5) +
    draw_plot(evolvability_plot, 0.5, 0, 0.5, 0.5) +
    draw_plot_label(c("A", "B", "C", "D"), c(0, 0.5, 0, 0.5),
                    c(1, 1, 0.5, 0.5), size = 20)
}
