% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{G_stats}
\alias{G_stats}
\title{Posterior draws for G matrix evolutionary statistics}
\format{list}
\source{
Data from: The evolution of phenotypic integration: how directional selection reshapes covariation in mice. \href{http://dx.doi.org/10.5061/dryad.5gr8r}{Dryad Archive} doi:10.5061/dryad.5gr8r.
}
\usage{
data(G_stats)
}
\description{
G matrix evolutionary statistics for all lines.
}
\examples{
data(G_stats)
\dontrun{
# To generate these from the posterior draws, run:
# (this can take a while)
# library(doMC)
# registerDoMC(5)
# G_stats = plyr::ldply(ratones_models, function(x) plyr::adply(x$Gs, 1,
# MeanMatrixStatistics, .progress = "text"), .parallel = TRUE)
# names(G_stats) <- gsub("pc1\%", "pc1.percent", names(G_stats))
}
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})

Melo, D., G. Garcia, A. Hubbe, A. P. Assis, and G. Marroig. 2016. "EvolQG - An R Package for Evolutionary Quantitative Genetics [version 3; Referees: 2 Approved, 1 Approved with Reservations]." F1000Research 4: 925.

Hansen, T. F., and Houle, D. (2008). Measuring and comparing evolvability and constraint in multivariate characters. Journal of evolutionary biology, 21(5), 1201-19. doi:10.1111/j.1420-9101.2008.01573.x
}
\keyword{datasets}
