% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generataAnimalModelInput.R
\name{generateAnimalModelInput}
\alias{generateAnimalModelInput}
\title{Generate input for BSFG model in Matlab}
\usage{
generateAnimalModelInput(formula, data, A, line = c("all", "t", "h", "hp",
  "s", "sp"), IDcol = "ID", out_folder = NULL)
}
\arguments{
\item{formula}{formula specifing fixed effects and response variables}

\item{data}{data.frame with data, default is ratonesdf}

\item{A}{relatinship matrix from pedigree}

\item{line}{line for matrix estimation. defaults to "all" for within-groups model}

\item{IDcol}{column with individual Ids matching the A matrix rownames}

\item{out_folder}{folder to create the Matlab input files}
}
\value{
list with animal model input objects
}
\description{
Generate input for BSFG model in Matlab
}
\examples{
data(ratonesdf)
data(ratones_ped)
formula = paste0("cbind(",
          paste(names(dplyr::select(ratonesdf, IS_PM:BA_OPI)),
                collapse = ", "), ") ~ SEX + AGE")
animal_model_data = generateAnimalModelInput(formula, ratonesdf,
                                             ratones_ped$A,
                                             line = "all", out_foldel = "./")
}
