% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{input vector}

\item{na.rm}{logical. If true, NA are removed before calculation}

\item{zero.propagete}{logical. If TRUE, checks the vector for zeros and skips calculation}
}
\value{
returns the geometric mean
}
\description{
Calculates the geometric mean of a vector
}
\examples{
x = runif(10, 1, 5)
gm_mean(x)
}
