% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{P_stats}
\alias{P_stats}
\title{Posterior draws for P matrix statistics}
\format{list}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(P_stats)
}
\description{
P matrix statistics for all lines.
}
\examples{
data(P_stats)
\dontrun{
# To generate these from the posterior draws, run:
# (this can take a while)
# library(doMC)
# registerDoMC(5)
# P_stats = plyr::ldply(ratones_models, function(x) plyr::adply(x$Ps, 1,
# MeanMatrixStatistics, .progress = "text"), .parallel = TRUE)
# names(P_stats) <- gsub("pc1\%", "pc1.percent", names(P_stats))
}
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
