% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{package}
\name{ratones-package}
\alias{ratones-package}
\title{Ratones package}
\description{
Companion package for the "The Evolution of Phenotypic Integration: How directional selection reshapes covariation in mice" manuscript.
}
