% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones}
\alias{ratones}
\title{Data for five mouse lines}
\format{List}
\source{
Data from: The evolution of phenotypic integration: how directional selection reshapes covariation in mice. \href{http://dx.doi.org/10.5061/dryad.5gr8r}{Dryad Archive} doi:10.5061/dryad.5gr8r.
}
\usage{
data(ratones)
}
\description{
Selection regime for body weight at 49 days was conducted in s and h lines for reduction, in s' and h' lines for increase, and a control line was maintained with random mating along generations, but avoiding full-sib matings. In downwards s and upwards s' lines inbreeding was performed by limiting population size and in downwards h and upwards h' full-sib mating was performed only during first generations.
}
\details{
List object organized by line, containing individual information and data.

\itemize{
 \item info.raw: complete information for all measured individuals, including both replicas.
         \itemize{
          \item ID: individual catalog number,
          \item SEX: sex (F: female, M: male),
          \item GER: generation of the long term experiment,
          \item LIN: experimental line (control t, downwards s, downwards h, upwards sp and upwards hp),
          \item MADRE: pedigree Dam,
          \item PADRE: pedigree Sire,
          \item AGE: age at sacrifice,
          \item P49: weight at 49 days of age,
          \item line: experimental line (control t, downwards s, downwards h, upwards s' and upwards h'),
          \item TAKE: paired individual replicas,
          \item selection: direction of selection (downwards, upwards or control),
          \item IS_PM-BA_OPI: set of 35 cranial euclidean distances, as indicated in the manuscript figure S3.
           }
 \item ed.raw: complete set of euclidean distances for all measured individuals, including replicas.
 \item info: information for all measured individuals.
 \item ed: individual mean between paired replicas of the 35 cranial euclidean distances.
 \item reps: vector of repeatabilities for the 35 cranial euclidean distances.
 \item model: linear model contoling for fixed effects (sex, age and generation).
 \item ed.means: vector of the line mean cranial euclidean distances.
 \item full: individual information, data and cranial euclinian distances'means between replicas.
\item gm_mean: line cranial geometric mean.
}
}
\examples{
data(ratones)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
