% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones_models}
\alias{ratones_models}
\title{Posterior draws for BSFG models}
\format{list}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(ratones_models)
}
\description{
G and P matrices for the skull distances for all 5 lines obtained from the BSFG model.
The generateAnimalModelInput function will create the inputs for the MATLAB code available in
http://www2.stat.duke.edu/~sayan/bfgr/index.shtml it you wish to generate these yourself.
}
\examples{
data(ratones_models)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
