% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones_ped}
\alias{ratones_ped}
\title{Pedigree information for five mouse lines}
\format{list}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(ratones_ped)
}
\description{
List with the pedigree for full experiment and relationship matrix for all measured animals
}
\examples{
data(ratones_ped)
if(!require(pedRSP)) devtools::install_github("jraffa/pedRSP");
library(pedRSP)
rsp = computeRSPs(kmat = ratones_ped$A, thre = 0.5)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
