% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones_Wmodel}
\alias{ratones_Wmodel}
\title{Posterior draws for BSFG within-groups model, including all 5 lines}
\format{list}
\source{
Data from: The evolution of phenotypic integration: how directional selection reshapes covariation in mice. \href{http://dx.doi.org/10.5061/dryad.5gr8r}{Dryad Archive} doi:10.5061/dryad.5gr8r.
}
\usage{
data(ratones_Wmodel)
}
\description{
G and P matrices for the skull distances joining all 5 lines in single model for within-group matrices.
The generateAnimalModelInput function will create the inputs for the MATLAB code available in
http://www2.stat.duke.edu/~sayan/bfgr/index.shtml it you wish to generate these yourself.
}
\examples{
data(ratones_Wmodel)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
