% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratonesdf}
\alias{ratonesdf}
\title{Linear distances for five mouse lines}
\format{data.frame}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(ratonesdf)
}
\description{
Skull distances measured from landmarks in 5 lines.
}
\examples{
data(ratonesdf)
delta_Z = colMeans(dplyr::select(ratonesdf[ratonesdf$selection == "upwards",], IS_PM:BA_OPI)) -
          colMeans(dplyr::select(ratonesdf[ratonesdf$selection == "downwards",], IS_PM:BA_OPI))
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
