% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorCor.R
\name{vectorCor}
\alias{vectorCor}
\title{Vector correlation}
\usage{
vectorCor(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Calculates the cosine correlation between two vectors
}
\examples{
x = rnorm(10)
y = rnorm(10)
vectorCor(x, y)
}
