% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones}
\alias{ratones}
\title{Linear distances for five mouse lines}
\format{List}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(ratones)
}
\description{
Skull distances measured from landmarks in 5 lines.
}
\examples{
data(ratones)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
