% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{package}
\name{ratones}
\alias{ratones}
\alias{ratones-package}
\alias{ratones}
\title{Ratones}
\format{List}
\source{
\href{http://dryad_link.com}{Dryad Archive}
}
\usage{
data(ratones)
}
\description{
Companion package for the Evolt... manuscript.

Skull distances measured from landmarks in 5 lines.
}
\examples{
data(ratones)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
