% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratones_ped}
\alias{ratones_ped}
\title{Pedigree information for five mouse lines}
\format{List}
\source{
Data from: The evolution of phenotypic integration: how directional selection reshapes covariation in mice. \href{http://dx.doi.org/10.5061/dryad.5gr8r}{Dryad Archive} doi:10.5061/dryad.5gr8r.
}
\usage{
data(ratones_ped)
}
\description{
List with the pedigree for full experiment and relationship matrix for all measured animals.
}
\details{
\itemize{
          \item ID: individual catalog number,
          \item MADRE: pedigree Dam,
          \item PADRE: pedigree Sire.
          }
}
\examples{
data(ratones_ped)
if(!require(pedRSP)) devtools::install_github("jraffa/pedRSP");
library(pedRSP)
rsp = computeRSPs(kmat = ratones_ped$A, thre = 0.5)
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
