% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratones-data.R
\docType{data}
\name{ratonesdf}
\alias{ratonesdf}
\title{Information and cranial euclidean distances for five mouse lines}
\format{data.frame}
\source{
Data from: The evolution of phenotypic integration: how directional selection reshapes covariation in mice. \href{http://dx.doi.org/10.5061/dryad.5gr8r}{Dryad Archive} doi:10.5061/dryad.5gr8r.
}
\usage{
data(ratonesdf)
}
\description{
Database for all individuals information and cranial euclidean distances mean between replicas.
}
\details{
\itemize{
          \item .id: selection and line combined information.
          \item ID: individual catalog number,
          \item SEX: sex (F: female, M: male),
          \item GER: generation of the long term experiment,
          \item LIN: experimental line (control t, downwards s, downwards h, upwards sp and upwards hp),
          \item MADRE: pedigree Dam,
          \item PADRE: pedigree Sire,
          \item AGE: age at sacrifice,
          \item P49: weight at 49 days of age,
          \item line: experimental line,
          \item TAKE: paired individual replicas (1 and 2),
          \item selection: direction of selection (downwards, upwards or control),
          \item IS_PM-BA_OPI: set of 35 cranial euclidean distances, as indicated in the manuscript figure S3.
           }
}
\examples{
data(ratonesdf)
delta_Z = colMeans(dplyr::select(ratonesdf[ratonesdf$selection == "upwards",], IS_PM:BA_OPI)) -
          colMeans(dplyr::select(ratonesdf[ratonesdf$selection == "downwards",], IS_PM:BA_OPI))
}
\references{
Penna et al. (2017) Evolution
(\href{http://www.ncbi.nlm.nih.gov/pubmed/**********}{PubMed})
}
\keyword{datasets}
